[feat] add validate password util
